doom upgrade: delete straight if recipe changed
By deleting straight, the latest pinned version will be reinstalled on
the second part of the 'doom upgrade' process.

This is a clumsy fix for a catch 22 where using straight to update
straight would break due to backwards incompatibilities (see #4950).
This will do while we wait for the CLI rewrite.