Prioritize emacs-mirror over melpa/elpa/etc
We can't be certain where MELPA/ELPA packages come from, but
occasionally, their hosted on unstable servers (savannah) or git servers
with certain features disabled (as is the case for paredit's not
allowing shallow clones).

Trust in emacsmirror, and we can be more certain the package is coming
from github.

This may trigger a rebuild of your packages!