add multi-account switching in notmuch module
introduces a new config +notmuch-mail-folders which holds a list of mail
folders. The +notmuch-mail-folder config that already exists is used as
the default. Nothing changes with normal usage of +notmuch/update for
syncing email folders, but it now accepts a prefix argument. When
supplied the prefix argument allows selection of a mail folder. This is
a non breaking change, adding a new feature that helps users with
multiple email folders.