fix(evil): +evil--embrace-escaped isn't active for the first buffer
For the first buffer by which `evil-embrace` is loaded,
`+evil--embrace-escaped` isn't applied. That is, surrounding text
object, say `blah` with backslash just result to `\blah\`, contrary to
waiting the very next character triggered by `+evil--embrace-escaped`.

This is because the entry associated with `\` doesn't exist when it
comes to the default value of `embrace--pairs-list`, hence, from the
very *next* buffer the tweak made to `embrace--pairs-list` takes effect.