fix(lookup): xref 1.3.0 removed var xref--marker-ring
Xref started many backwards incompatible changes in the name of
performance.  The marker ring no longer exists as a variable.
+lookup--xref-show will silently fail every time.

Switches to using xref-after-jump-hook to test whether xref actually
jumped to a definition. The hook is part of the exposed interface and
should hopefully not change in the near future.

Fix: #5737