fix(default): search symbol at point from project root
...Rather than current directory.

While +default/search-project-for-symbol-at-point should operate out of
the current project's root, +default/search-notes-for-symbol-at-point
must operate out of org-directory, and no where else, even if
org-directory is not a valid project. 411ed0c8f8e1 was written to fix
that, but it indirectly broke
'+default/search-project-for-symbol-at-point' (causing it to always
resolve to the current directory, rather than the current project's
root).

Amend: 411ed0c8f8e1
Fix: #5831