fix: skip deploy git hooks when doom dir is not a repository.
In the nix-doom-emacs, the doom dir is not a repository, in this case,
`git rev-parse --git-path hook` returns an error,
and doom-emacs tries to create a new directory using error text as
directory name.
In nixos it cause permission error due to immutable file system, in
other systems it could create a wrong directory.