refactor!(erlang): remove company-erlang
BREAKING CHANGE: Remove `company-erlang` which depends on
`ivy-erlang-complete`. `company-erlang` and `ivy-erlang-complte` have
not been maintained now. The lsp mode provides much better experience
than `company-erlang` and `ivy-erlang-complete`. The
`ivy-erlang-complete` package also has the defect that slow down the
erlang-mode buffer when the buffer is huge enough. It is not a good
choice to integrate ivy-erlang-complete nowadays even as a non-LSP
option.