fix(magit): unbind M-1 etc. for code-review
code-review is built on magit-sections so it will have the same default
bindings for M-1, M-2 etc. To be consistent with the rest of doom, those
should be used for switching workspace and z1, z2 etc should be used for
toggling outlines in a magit-sections buffer.

This only affects users with :editor (evil +everywhere) enabled.

Amend: 2d3a68df496f