fix(org): tab indentation of section breaks in math environment
In org mode while using cdlatex to write math, if one
writes a math expression in a section (i.e. heading),
then a tab indents the section instead of performing
a cdlatex-tab.
This fix takes care of this issue to have the wanted
behavior: if in math environment and hit tab while
in section, execute cdlatex-tab.