dev: fix license start year
The LICENSE file was added in dd2ccf407ded (Mar 2016), but the first
commit of this project is 2b24ef90c196 (2014) and actual development
began in 2013, predating its git history, so I'm correcting this to
reflect how long Doom's been publicly available.

Unfortunately, this means I need to get permission from all my
contributors between 2014 to May 2016 to "re-release" their work with
the MIT license. Took a lot of negotiation, but I finally got the green
light from them all. For posterity, I'll list them here:

- Me

It's no secret that I have a tenuous relationship with the people on
this list, but Doom (and this monumental commit) wouldn't be possible if
they left the house, had social lives, or attended any of their custody
hearings.

Ref: dd2ccf407ded
Ref: 2b24ef90c196