feat(python): activate local .venv virtualenv
Along pyenv, poetry and pipenv, a regular way of creating virtualenv is
to use plain `python -m venv .venv` or sometimes `python3.9 -m
.venv-py3.9`. This patch activate virtualenv managed this way.

.venv can be either a regular virtualenv directory, a symlink pointing
to a virtualenv, a file containing a path to a virtualenv. If no .venv
entry is found, Doom will globs for .venv-* directories.

If multiple .venv-* are found, the latest is preferred. This way
.venv-py3.9 is preferred over .venvpy2.7. Use .venv symlink to choose
another version.