docs(spell): can't add word to personal dict
Adding `Cannot add [word] to any active dictionary` error to the
troubleshooting section.

I have had this problem multiple times now when I break my emacs and
wipe the `~/.config/emacs` directory and I can never remember what file
to create to get my spell check working properly.

Even though this is a bug and might be fixed in the future, I feel like
having it documented can be helpful in the short term and in the future.

References: #6246