fix(lib): unquote dir! call in load! macro
This restores previous behavior where dir! was called at the time of
load!'s expansion, which ensured that it resolved to the directory of
the source file where load! was called

In my case, this manifested in not being able to use evil-ex because of
a "Cannot get this file-path" error _unless_ evil-ex was called with
modules/editor/evil/config.el in the active buffer first to load it