feat(workspaces): add mru cache for workspaces
This enables keeping a list of workspaces in most recently used order so
that we can enable "alt-tab"-like functionality for switching between
workspaces that are not apps (or vice versa, only switching between
apps).

This is currently behind an +mru flag to workspaces but it doesn't
necessarily need to be since it doesn't change any behavior. Users would
need to opt-in to using the new functions.

Need to decide other keybindings for evil, if any.
Also, need to decide if this should be a flag or not.