dev: remove feature/bump request issue forms
This removes the feature and bump request templates, as our issue
tracker will no longer accept them. They should either be submitted to
our Discourse/Discord, or come in the form of a pull request (as a proof
of concept proposal/RFC), or a bug report (in case a bump is needed to
address a bug).

Ref: https://discourse.doomemacs.org
Ref: https://doomemacs.org/discord