fix: doom/delete-this-file nil path handling
If path is nil (e.g. called interactively when buffer is not visiting a
file), avoid calling abbreviate-file-name on nil, otherwise an error
will be signalled:

  (wrong-type-argument stringp nil)

Additionally, fixup the subsequent path checks. These were treating two
distinct scenarios as a "Buffer is not visiting any file" user-error:
  - nil path
  - non-existent path

Only the first should result in that error. The second should proceed to
the next path check (which was previously unreachable), to signal the
appropriate error, "File doesn't exist: %s".