fix(lsp): don't prompt twice when selecting root
When the lsp root can't be computed automatically (via
`lsp--calculate-root`), lsp-mode will prompt the user to select it.

Doom's logging hook was calling `lsp--calculate-root` again to know what
to log (assuming it would just compute it again based on directory
structure), which means we would sometimes as the user to select the
root twice.

Here we reimplement the logging using advice instead so we can piggyback
on the first call and log whatever it returns.