Update treemacs pin to fix `dap-mode`
`treemacs` updated the interface to `treemacs-initialize` recently:

https://github.com/Alexander-Miller/treemacs/commit/20765acd38e00faa46a72b9a2cf63a7b451c6850#diff-36f7ea49dcbb6f011b9faf5f012271328ac46515b4f161214001917cbc2ef6b2L1023-R1053

The currently pinned version of `lsp-treemacs` expects the new interface, but treemacs is pinned to an older commit, before the interface was changed. This causes `dap-mode` to error out when launching the UI debugger:

https://github.com/emacs-lsp/lsp-treemacs/issues/145#issuecomment-1274966247

Update the pin so `treemacs` is serving the new interface, plus some other additions.