fix(lsp): only log computed root if using lsp-auto-guess-root
Only logging the computed root when using lsp-auto-guess-root means we
don't risk logging a root that might not match the user's manual
selection.

This is mainly to avoid relying on `defadvice!` and peeking into
lsp-mode's internals to figure out the actually selected root for the
buffer (assuming that's even possible).