fix(email): mu4e main view shortcut face
In the leuven theme (and possibly others), the use of
`(mode-line-emphasis bold)` results in white text on a white background,
making the shortcut keys invisible.

The original function mu4e--main-action-string uses mu4e-highlight-face,
and changing back to this fixes the problem.

I can’t tell from the git history what the reasoning for the face change
specifically was: the new function seems to have been made mainly for
changing the bullets to unicode versions.

Fix: #7010