Correctly `ln` the `Emacs.app` after homebrew install
Output from a freshly installed homebrew, on a brand new Mac: 

>>> 

This is YAMAMOTO Mitsuharu's "Mac port" addition to
GNU Emacs 28. This provides a native GUI support for Mac OS X
10.10 - 12. After installing, see README-mac and NEWS-mac
in /opt/homebrew/opt/emacs-mac for the port details.

Emacs.app was installed to:
  /opt/homebrew/opt/emacs-mac
>>>


When I was following the instructions, I wasn't able to execute Emacs afterwards. With this change to the `ln` I'm able to run `Emacs.app`